Port CI setup from Travis to GitHub Actions (#150)
* Port CI setup from Travis to GitHub Actions

The CI setup on Travis CI stopped working. This commit implements the
same setup on GitHub Actions (additionally adding Ruby 3.x to the test
matrix). The old Travis CI setup was removed.

* Fix Rubocop on Ruby > 3.0

Rubocop 0.48 does not support Ruby 3.1 and later, so on those versions
require rubocop ~> 1.45. This also introduces many new cops. For now,
disable those new coops to keep parity with the previous setup. Ideally,
they should be manually checked, to decide whether to enable the cop and
fix the violation, or keep the cop disabled.