Handle instantaneous and unsolicited PUBACKs (#158)
Prior to this commit, it was possible to crash the read thread
by sending a PUBACK with an unexpected ID, e.g. an ID the client
had not used before, or had already deleted the queue for.
These packets will now simple be ignored.

Furthermore, with QoS 1 or 2, the PUBACK queue for a packet
was only created _after_ the packet had already been published.
If the PUBACK arrived within this tiny gap, the read thread would crash.
We fix this by creating the queue before starting to publish.